\name{act.scores.06.07}
\alias{act.scores.06.07}
\title{ACT scores of states in the US.}
\description{
Average ACT composite scores for all states for the 2006-07 year.
}
\usage{
act.scores.06.07
}
\format{
  \item{STATE}{name of state}
  \item{Abbrev}{state abbreviation}
  \item{ACT}{average ACT composite score}
  \item{Per.Grads}{percentage of students who took the ACT exam}
  \item{Region}{region of the United States}
}

\keyword{datasets}