\name{baseball.attendance}
\alias{baseball.attendance}
\docType{data}
\title{Attendance at baseball teams.}
\description{
Average attendance for home and away games for all major league
baseball games in the 2001 season.
}
\usage{
baseball.attendance
}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
  \item{Team}{name of team}
  \item{League}{league where the team belongs}
  \item{Home.Attendance}{average attendance for home games}
  \item{Away.Attendance}{average attendance for away games}
}
}
\source{www.espn.com web site}
\keyword{datasets}
