\name{beatles}
\alias{beatles}
\title{Lengths of songs from Beatles' albums.}
\description{
Length of songs from a selection of albums from the popular
pop and rock group The Beatles.
}
\usage{
beatles
}
\format{
  \item{time}{state abbreviation}
  \item{album}{name of Beatles' album}
}

\keyword{datasets}
