\name{boston.marathon}
\alias{boston.marathon}
\title{Boston Marathon completion times of women of different ages.}
\description{
Completion times for a sample of women at the 2001 Boston Marathon.
}
\usage{
boston.marathon
}
\format{
  \item{time}{completion time in minutes}
  \item{age}{age of the participant}
}

\keyword{datasets}
