\name{boston.marathon.wtimes}
\alias{boston.marathon.wtimes}
\title{Boston marathon winning times}
\description{
Winning time in mens Boston marathon through the years.
}
\usage{
boston.marathon.wtimes
}
\format{
  \item{year}{year}
  \item{minutes}{winning time in minutes}
}

\keyword{datasets}

