\name{brave.at}
\alias{brave.at}
\docType{data}
\title{Attendance at games of a baseball team}
\description{
Attendance of Atlanta Braves home games during a particular season.
}
\usage{
brave.at
}
\format{
  A data frame with 72 observations on the following 2 variables.
  \describe{
  \item{attend}{attendance count}
  \item{game}{game number}
  }
}
\source{www.espn.com website}
\keyword{datasets}

