\name{car.measurements}
\alias{car.measurements}
\title{Car measurements}
\description{
Measurements on 38 automobiles from the 1978-79 model year.
}
\usage{
car.measurements
}
\format{
  \item{Country}{the nationality of the manufacturer}
  \item{Car}{the car name}
  \item{MPG}{the mileage, measured in miles per gallon (MPG)}
  \item{Weight}{the weight}
  \item{Drive.Ratio}{the drive ratio}
  \item{Horsepower}{the horsepower}
  \item{Displacement}{the displacement 
of the car (in cubic inches)}
  \item{Cylinders}{the number of cylinders}
}

\keyword{datasets}

	