\name{church.2way}
\alias{church.2way}
\title{Church attendance as a two-way table.}
\description{
Two-way table of average worship attendance at a Ohio church by year and  month.
}
\usage{
church.2way
}
\format{
  \item{month}{month}
  \item{y1993}{average worship attendance count in 1993}
  \item{y1994}{average worship attendance count in 1994}
  \item{y1995}{average worship attendance count in 1995}
  \item{y1995}{average worship attendance count in 1996}
 }

\keyword{datasets}
