\name{church.tseries}
\alias{church.tseries}
\title{Time series of worship attendance}
\description{
Time series of weekly worship attendance from a church in Ohio, U.S.
}
\usage{
church.tseries
}
\format{
  \item{year}{the year}
  \item{month}{the month}
  \item{worship}{count of worship attendance}
 }

\keyword{datasets}