\name{farms}
\alias{farms}
\title{Number of farms in the states of the U.S.}
\description{
The number of farms (in 1000s) for each of the 50 states 
in the United States in 1999.
}
\usage{
farms
}
\format{
  \item{state}{name of the state}
  \item{count}{count of the number of farms}
}

\keyword{datasets}
