\name{gestation.periods}
\alias{gestation.periods}
\title{Gestation periods for different animals.}
\description{
Average gestation periods for different animals.
}
\usage{
baseball.attendance
}
\format{
  \item{Animal}{name of animal}
  \item{Period}{average gestation period}
}

\keyword{datasets}
