\name{grandma.19.40}
\alias{grandma.19.40}
\title{Grandmas Marathon completion times}
\description{
Completion times in running Grandmas Marathon (Duluth, Minnesota) 
for women of ages 19 to 34}

\usage{
grandma.19.40
}
\format{
  \item{time}{completion time in seconds}
    }

\keyword{datasets}
