\name{home.prices}
\alias{home.prices}
\title{Home sales prices in the U.S.}
\description{
Average sales prices of houses for 21 cities in the U.S. 
}
\usage{
home.prices
}
\format{
  \item{City}{name of the city}
  \item{y1985}{average home price in 1985}
  \item{y2000}{average home price in 2000}
 }

\keyword{datasets}
