\name{homeruns.2000}
\alias{homeruns.2000}
\title{Team home run numbers for different seasons.}
\description{
Total number of home runs hit all teams in the major league baseball
league for the years 1900,1910,..., 2000.
}
\usage{
homeruns.2000
}
\format{
  \item{YEARS}{season}
  \item{HOMERUNS}{count of team home runs}
}

\keyword{datasets}
