\name{immigrants}
\alias{immigrants}
\docType{data}
\title{Immigrant counts to US.}
\description{
Immigrant counts to US from various countries.
}
\usage{
immigrants
}
\format{
  A data frame with 41 observations on the following 3 variables.
  \describe{
  \item{Country}{name of country}
  \item{Region}{region of the world}
  \item{X1998.Count}{the count of immigrants in 1998}
  }
}
\source{
2001 New York Times Almanac
}
\keyword{datasets}
