\name{island.areas}
\alias{island.areas}
\title{Areas of islands from different continents.}
\description{
Total number of home runs hit all teams in the major league baseball
league for the years 1900,1910,..., 2000.
}
\usage{
islands.areas
}
\format{
  \item{Ocean}{ocean where the island lies}
  \item{Name}{name of the island}
  \item{Area}{area of the island}
}

\keyword{datasets}
