\name{lake}
\alias{lake}
\title{Lake measurements}
\description{
Measurements of lakes in the Vilas and Oneida counties 
of northern Wisconsin (from Minitab dataset collection).
}
\usage{
lake
}
\format{
  \item{Area}{area of lake in acres}
  \item{Depth}{maximum depth of lake in feet}
  \item{PH}{pH (acidity) measurement}
  \item{Wshed}{watershed area in square miles}
  \item{Hions}{concentration of hydrogen ions}
}

\keyword{datasets}

