\name{mortality.rates}
\alias{mortality.rates}
\docType{data}
\title{Infant mortality rates of countries}
\description{
Infant mortality rates in 1997 for 68 countries (Afghanistan through Guyana).
}
\usage{
mortality.rates
}
\format{
  A data frame with 68 observations on the following 1 variable.
  \describe{
  \item{rate}{infant mortality rate}
  }
}
\source{2001 New York Times Almanac}
\keyword{datasets}

