\name{olympics.run}
\alias{olympics.run}
\title{Olympics running times.}
\description{
Winning time in Olympics mens running race classified by length of race and year.
}
\usage{
olympics.run
}
\format{
  \item{year}{year of the Olympics}
  \item{100m}{winning time in seconds for 100m}
  \item{200m}{winning time in seconds for 200m}
  \item{400m}{winning time in seconds for 400m}
  \item{800m}{winning time in seconds for 800m}
  \item{1500m}{winning time in seconds for 1500m}
  }

\keyword{datasets}
