\name{olympics.speed.skating}
\alias{olympics.speed.skating}
\title{Olympics speed skating times.}
\description{
Winning time in Olympics mens speed skating race classified by length of race and year.
}
\usage{
olympics.speed.skating
}
\format{
  \item{year}{year of the Olympics}
  \item{500m}{winning time in seconds for 500m}
  \item{1000m}{winning time in seconds for 1000m}
  \item{1500m}{winning time in seconds for 1500m}
  \item{5000m}{winning time in seconds fro 5000m}
  \item{10000m}{winning time in seconds for 10000m}
  }

\keyword{datasets}