\name{olympics.swim}
\alias{olympics.swim}
\title{Olympics swimming times.}
\description{
Winning time in Olympics womens swimming race classified by length  of race and year.
}
\usage{
olympics.swim
}
\format{
  \item{year}{year of the Olympics}
  \item{100m}{winning time in seconds for 100m}
  \item{200m}{winning time in seconds for 200m}
  \item{400m}{winning time in seconds for 400m}
  \item{800m}{winning time in seconds for 800m}
  }

\keyword{datasets}
