\name{pop.change}
\alias{pop.change}
\docType{data}
\title{Population change for all states in the U.S.}
\description{
Population change for all states in the United States between the
years 1900 and 1999.
}
\usage{
pop.change
}
\format{
  A data frame with 51 observations on the following 2 variables.
  \describe{
  \item{State}{state abbreviation}
  \item{Pct.change}{percentage change in population}
  }
}
\source{World Almanac and Book of Facts 2001}
\keyword{datasets}
