\name{rent.prices}
\alias{rent.prices}
\title{Rent prices in different cities.}
\description{
Rental prices (different types from different cities).}

\usage{
rent.prices
}
\format{
  \item{City}{city}
  \item{Studio}{rent of a studio apartment}
  \item{One.Bedroom}{rent of a one bedroom apartment}
  \item{Two.Bedroom}{rent of a two bedroom apartment}
  \item{Three.Bedroom}{rent of a three bedroom apartment}
  \item{Four.Bedroom}{rent of a four bedroom apartment}
  }

\keyword{datasets}
