\name{salaries}
\alias{salaries}
\title{Salaries of different professions in different cities.}
\description{
1979 salaries of seven different professions in six cities.
}
\usage{
salaries
}
\format{
  \item{Salary}{salary in hundreds of Swiss francs}
  \item{City}{name of city}
  \item{Profession}{name of profession}
}

\keyword{datasets}