\name{temperatures}
\alias{temperatures}
\title{Temperatures for different cities.}
\description{
Two-table of average temperatures of cities for different months.}

\usage{
temperatures
}
\format{
  \item{City}{city}
  \item{January}{average temperature in January}
  \item{April}{average temperature in April}
  \item{July}{average temperature in July}
  \item{October}{average temperature in October}
  }

\keyword{datasets}
