\name{immigrants}
\alias{immigrants}
\title{Immigrant counts to US.}
\description{
Immigrant counts to US from various countries.
}
\usage{
immigrants
}
\format{
  \item{Country}{name of country}
  \item{Region}{region of the world}
  \item{1998.Count}{the count of immigrants in 1998}
}

\keyword{datasets}
