\name{pop.change}
\alias{pop.change}
\title{Population change for all states in the U.S.}
\description{
Population change for all states in the United States between the
years 1900 and 1999.
}
\usage{
pop.change
}
\format{
  \item{State}{state abbreviation}
  \item{Pct.change}{percentage change in population}
}

\keyword{datasets}
