\name{pop.densities.1920.2000}
\alias{pop.densities.1920.2000}
\docType{data}
\title{Population densities of states for different years.}
\description{
Population density of each state in the United States for the years
1920, 1960, 1980, 1990, 2000.
}
\usage{
pop.densities.1920.2000
}
\format{
  A data frame with 51 observations on the following 6 variables.
  \describe{
  \item{State}{names of state}
  \item{y1920}{population density of state in 1920}
  \item{y1960}{population density of state in 1960}
  \item{y1980}{population density of state in 1980}
  \item{y1990}{population density of state in 1990}
  \item{y2000}{population density of state in 2000}
  }
}
\source{World Almanac and Book of Facts 2001}
\keyword{datasets}
